Redistributing code to ReadImageInformation().
The ReadImageInformation() method is supposed to get to the point
where it figures out the size of the input image.

Unfortunately, it seems that we need to decode the VP8 frame in
order to grab that information. (There is probably a better way
to do it...).  Therefore, the initial part of the decoding is
now done in ReadImageInformation() and then the reading continues
in Read().  We are not managing multiple components yet, nor we
have a clear plan for what to do when the image has multiple frames
in a single file.